Search Venues from Telegram
Disable error so Telegram doesn't retry endlessly
Location Parameter fix
Changed Business Venues to actually return objects
Replaced logic from the ListsController to service and models.
Security by logged status
Available messages Keyboard
Markdown on messages

+++ app/models/telegram/Location.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Location {
 
-    public Float longitude;
-    public Float latitude;
+    public double longitude;
+    public double latitude;
 
 }
